Move pin wrappers to src/pins
