+++ speeduino/scheduledIO_direct_inj.cpp
@@ -1,6 +1,6 @@
 #include "globals.h"
 #include "board_definition.h"
-#include "fastOutputPin.h"
+#include "src/pins/fastOutputPin.h"
 #include "preprocessor.h"
 
 // LCOV_EXCL_START
